multi-agent-docker/requirements.txt
wgpu==0.22.2
tensorflow
keras
h2o
xgboost
flake8
pylint
google-generativeai
beautifulsoup4
aiohttp
# New 2D/3D Additions
# open3d - temporarily disabled, no compatible wheel for Python 3.12
pymeshlab
# Note: nerfstudio requires PyAV which has Cython compilation issues
# If you need nerfstudio, install it separately after the base image is built:
# pip install nerfstudio
# For now, commenting it out to ensure the build succeeds
# nerfstudio
# Note: gaussian-splatting-pytorch is not a standard PyPI package name.
# A specific implementation will need to be installed via pip git+https,
# which can be added later once a repository is chosen.
# For now, this is a placeholder.
# gaussian-splatting-pytorch